Fixed issue with port where it was not cleaning up
Ports that had empty eta and etd’s were not being removed. Fix was to
create a copy of the original array by adding a [:] to the end of the
loop iterating over the ports array. This created a copy so that when
we removed it was not editing a list while iterating.